fix: 禁止 Full 模式修改 Markdown 原文
